auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["gahara <madinikus@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "8.1.3"
Flask = "2.2.2"
importlib-metadata = "5.0.0"
itsdangerous = "2.1.2"
Jinja2 = "3.1.2"
MarkupSafe = "2.1.1"
typing_extensions = "4.4.0"
Werkzeug = "2.2.2"
zipp = "3.8.1"
common = {path = "../common"}
flask-sqlalchemy = "^3.0.0"
Flask-Migrate = "^3.1.0"
PyJWT = "2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
